ci: run codecov outside of the image
First, by default the CI environment variables aren't passed into the
image, so codecov can't detect its environment. This can be fixed by
running the codecov.io/env script to obtain the docker run env vars

Second, codecov expects to have access to the git binary, which isn't
available in the alpine-based node image.

Instead we'll mount a volume inside the image where the coverage is
expected to appear, so that it's accessible outside of the image. Then
the generic codecov script is run to submit them.