feat: add Config class
This consolidates all configuration and environment variables into one
location so that it's easy to see at a glance, and enforce, the
configuration of the application, rather than having different parts of
the system implicitly depending on global configuration variables.

Config is a singleton which caches process.env on construction. Required
configuration values are then read into explicit instance variables with
default values, optionally lazily computed. Those without default values
are considered required and the application throws an error if they're
missing.