refactor: reverse De Morgan's Laws
De Morgan's Laws is a simple Boolean optimization/transformation that
the database surely can do on its own, and even if not, it would
probably be a micro-optimization, not to mention premature, even if
optimization was not my original intention.

Simplicity and readability should be paramount when concerning table
constraints, and transformation by De Morgan's Laws makes things more
difficult to reason about.

!(P && Q) === !P || !Q

https://en.wikipedia.org/wiki/De_Morgan%27s_laws