build: move sql files to ./db/sql/
Note that this requires setting the Docker build context to ./db/sql/
and the Dockerfile path specified relative to it.

This is mainly to allow the schema and seed data to live outside of the
Docker-specific directories, the rationale being that in production the
database may not be running within Docker (e.g. Google Cloud SQL).